refactor : 패키지 이름 변경 (단수형 -> 복수형)
