+++ user-service/src/main/java/com/team/userservice/user/companies/infrastructure/CompanyRepositoryImpl.java
@@ -1,8 +1,8 @@
-package com.team.userservice.user.company.infrastructure;
+package com.team.userservice.user.companies.infrastructure;
 
 import com.team.userservice.global.domain.error.UserErrorCode;
 import com.team.userservice.global.exception.UserException;
-import com.team.userservice.user.company.domain.CompanyRepository;
+import com.team.userservice.user.companies.domain.CompanyRepository;
 import com.team.userservice.user.core.CompanyUser;
 import com.team.userservice.user.core.User;
 import java.util.UUID;
