+++ user-service/src/main/java/com/team/userservice/user/core/CompanyUser.java
@@ -1,5 +1,6 @@
 package com.team.userservice.user.core;
 
+import com.team.common.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -12,6 +13,7 @@ import java.util.UUID;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.UuidGenerator;
+import com.team.common.BaseEntity;
 
 @Entity
 @Table(name = "p_company_user")
